# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/config": {
      "get": {
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfig"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patch_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfig"
                }
              }
            }
          }
        }
      }
    },
    "/config/providers/{id}/secret": {
      "put": {
        "summary": "Save (or overwrite) the keyring secret for a provider. Creates the\nprovider entry in `config.providers` if it didn't exist. `PUT` because\nsetting the secret is idempotent for the same body.",
        "operationId": "set_provider_secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSecretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      },
      "delete": {
        "summary": "Clear a provider's keyring secret. The provider entry itself is kept.",
        "operationId": "clear_provider_secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/operations/{id}": {
      "delete": {
        "operationId": "cancel_operation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Operation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/pipelines": {
      "post": {
        "operationId": "start_pipeline",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartPipelineResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppConfig": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DataConfig"
              }
            ],
            "default": {
              "path": "C:\\Users\\datpt02\\AppData\\Local\\Koharu"
            }
          },
          "http": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/HttpConfig"
              }
            ],
            "default": {
              "connect_timeout": 20,
              "max_retries": 3,
              "read_timeout": 300
            }
          },
          "pipeline": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PipelineConfig"
              }
            ],
            "default": {
              "bubble_segmenter": "speech-bubble-segmentation",
              "comic_text_bubble_detector_classes": [
                "text_bubble",
                "text_free"
              ],
              "detector": "pp-doclayout-v3",
              "detector_confidence_threshold": null,
              "font_detector": "yuzumarker-font-detection",
              "inpainter": "lama-manga",
              "ocr": "paddle-ocr-vl-1.6",
              "renderer": "koharu-renderer",
              "segmenter": "comic-text-detector-seg",
              "segmenter_binary_threshold": null,
              "translator": "llm",
              "unlimited_ocr_mode": "off",
              "unlimited_ocr_url": null
            }
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderConfig"
            },
            "default": []
          }
        }
      },
...
      "ConfigPatch": {
        "type": "object",
        "description": "Sparse patch for `koharu_app::AppConfig`. Missing fields mean \"leave\nas-is\". The `providers` field, if present, replaces the whole provider\nlist — we do not merge by id because ordering is meaningful.",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DataConfigPatch"
              }
            ]
          },
          "http": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HttpConfigPatch"
              }
            ]
          },
          "pipeline": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PipelineConfigPatch"
              }
            ]
          },
          "providers": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/ProviderPatch"
            },
            "description": "If present, replaces the entire list. Api_key values of `\"[REDACTED]\"`\nare interpreted as \"leave the existing secret alone\"."
          }
        }
      },
...
      "DataConfig": {
        "type": "object",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          }
        }
      },
      "DataConfigPatch": {
        "type": "object",
        "properties": {
          "path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "HttpConfig": {
        "type": "object",
        "properties": {
          "connect_timeout": {
            "type": "integer",
            "format": "int64",
            "default": 20,
            "minimum": 0
          },
          "max_retries": {
            "type": "integer",
            "format": "int32",
            "default": 3,
            "minimum": 0
          },
          "read_timeout": {
            "type": "integer",
            "format": "int64",
            "default": 300,
            "minimum": 0
          }
        }
      },
      "HttpConfigPatch": {
        "type": "object",
        "properties": {
          "connectTimeout": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "maxRetries": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "readTimeout": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "NodeId": {
        "type": "string",
        "format": "uuid"
      },
...
      "PageId": {
        "type": "string",
        "format": "uuid"
      },
...
      "PipelineConfig": {
        "type": "object",
        "description": "Engine selection for each pipeline stage.\nValues are engine IDs (e.g. \"pp-doclayout-v3\", \"comic-text-detector\").\nEmpty string means use default.",
        "properties": {
          "bubble_segmenter": {
            "type": "string",
            "default": "speech-bubble-segmentation"
          },
          "comic_text_bubble_detector_classes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": [
              "text_bubble",
              "text_free"
            ]
          },
          "detector": {
            "type": "string",
            "default": "pp-doclayout-v3"
          },
          "detector_confidence_threshold": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Confidence threshold override (0.0..=1.0) for the currently selected\n`detector` engine. `None` = use that engine's built-in default.",
            "default": null
          },
          "font_detector": {
            "type": "string",
            "default": "yuzumarker-font-detection"
          },
          "inpainter": {
            "type": "string",
            "default": "lama-manga"
          },
          "ocr": {
            "type": "string",
            "default": "paddle-ocr-vl-1.6"
          },
          "renderer": {
            "type": "string",
            "default": "koharu-renderer"
          },
          "segmenter": {
            "type": "string",
            "default": "comic-text-detector-seg"
          },
          "segmenter_binary_threshold": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Binary mask threshold override (0.0..=1.0) for the currently selected\n`segmenter` engine. `None` = use that engine's built-in default.",
            "default": null
          },
          "translator": {
            "type": "string",
            "default": "llm"
          },
          "unlimited_ocr_mode": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/UnlimitedOcrMode"
              }
            ],
            "default": "off"
          },
          "unlimited_ocr_url": {
            "type": [
              "string",
              "null"
            ],
            "default": null
          }
        }
      },
      "PipelineConfigPatch": {
        "type": "object",
        "properties": {
          "bubbleSegmenter": {
            "type": [
              "string",
              "null"
            ]
          },
          "comicTextBubbleDetectorClasses": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "detector": {
            "type": [
              "string",
              "null"
            ]
          },
          "detectorConfidenceThreshold": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "`Some(Some(x))` sets an override, `Some(None)` clears it back to the\ndetector's built-in default, `None` leaves the existing value as-is."
          },
          "fontDetector": {
            "type": [
              "string",
              "null"
            ]
          },
          "inpainter": {
            "type": [
              "string",
              "null"
            ]
          },
          "ocr": {
            "type": [
              "string",
              "null"
            ]
          },
          "renderer": {
            "type": [
              "string",
              "null"
            ]
          },
          "segmenter": {
            "type": [
              "string",
              "null"
            ]
          },
          "segmenterBinaryThreshold": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "`Some(Some(x))` sets an override, `Some(None)` clears it back to the\nsegmenter's built-in default, `None` leaves the existing value as-is."
          },
          "translator": {
            "type": [
              "string",
              "null"
            ]
          },
          "unlimitedOcrMode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnlimitedOcrMode"
              }
            ]
          },
          "unlimitedOcrUrl": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ProviderConfig": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "api_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "Populated from credential storage on `load()`, never written to config.toml.\nSerializes as `\"[REDACTED]\"` in API responses."
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "max_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        }
      },
      "ProviderPatch": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "apiKey": {
            "type": [
              "string",
              "null"
            ],
            "description": "`\"[REDACTED]\"` → keep existing keyring secret; empty → clear; otherwise save."
          },
          "baseUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "`None` = unchanged. `Some(None)` = clear. `Some(Some(x))` = set."
          },
          "id": {
            "type": "string"
          },
          "maxTokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "`None` = unchanged. `Some(None)` = clear. `Some(Some(x))` = set.",
            "minimum": 0
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "description": "`None` = unchanged. `Some(None)` = clear. `Some(Some(x))` = set."
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "`None` = unchanged. `Some(None)` = clear. `Some(Some(x))` = set."
          }
        }
      },
      "ProviderSecretRequest": {
        "type": "object",
        "required": [
          "secret"
        ],
        "properties": {
          "secret": {
            "type": "string"
          }
        }
      },
...
      "ReadingOrder": {
        "type": "string",
        "enum": [
          "rtl",
          "ltr",
          "custom"
        ]
      },
      "Region": {
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "properties": {
          "height": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "width": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "x": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "y": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "StartPipelineRequest": {
        "type": "object",
        "required": [
          "steps"
        ],
        "properties": {
          "defaultFont": {
            "type": [
              "string",
              "null"
            ]
          },
          "pages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/PageId"
            },
            "description": "`None` → whole project, `Some(pages)` → just those pages."
          },
          "readingOrder": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReadingOrder"
              }
            ]
          },
          "region": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Region",
                "description": "Optional bounding-box hint for inpainter engines (repair-brush)."
              }
            ]
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Engine ids (`inventory::submit!` ids) to run in order."
          },
          "systemPrompt": {
            "type": [
              "string",
              "null"
            ]
          },
          "targetLanguage": {
            "type": [
              "string",
              "null"
            ]
          },
          "textNodeIds": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/NodeId"
            },
            "description": "Optional text-node ids for engines that can operate on individual blocks."
          },
          "unlimitedOcrMode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnlimitedOcrMode",
                "description": "`None` = fall back to saved config."
              }
            ]
          },
          "unlimitedOcrUrl": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "StartPipelineResponse": {
        "type": "object",
        "required": [
          "operationId"
        ],
        "properties": {
          "operationId": {
            "type": "string"
          }
        }
      },
...
      "UnlimitedOcrMode": {
        "type": "string",
        "description": "Controls how the pipeline handles unlimited-OCR routing.",
        "enum": [
          "off",
          "full",
          "smart-fallback"
        ]
      }
    }
  }
}